+++ radix-cub.cuh
@@ -5,7 +5,7 @@
 #include"cub/cub.cuh"
 
 template<typename T>
-void RadixSortCub(T* d_in, T* d_out, int N) {
+void RadixSortCub(T* d_in, T* d_out, size_t N) {
     void* d_tmp_storage = NULL;
     size_t tmp_storage_bytes = 0;
     cub::DeviceRadixSort::SortKeys(d_tmp_storage, tmp_storage_bytes, d_in, d_out, N);
